Middleware Authorization

+++ app/Http/Controllers/JobController.php
@@ -49,13 +49,6 @@ class JobController extends Controller
 
     public function edit(Job $job)
     {
-        /**
-         * The User model that ships with laravel,
-         * includes access to "can" and "cannot" method.
-         * They are unique and specifically for authorization.
-         */
-        // if (Auth::user()->can('edit-job', $job)) { ... }
-
         /** 
          * Runs the logic (Gate::define(...)) associated with the name referenced "edit-job",
          * if it fails or return false, laravel automatically aborts with a 403.
@@ -64,7 +57,7 @@ class JobController extends Controller
          * if (Gate::allows(...)) {...}
          * if (Gate::denies(...)) {...}
          */
-        // Gate::authorize('edit-job', $job);
+        Gate::authorize('edit-job', $job);
 
         return view('jobs.edit', [
             'job' => $job
@@ -73,7 +66,7 @@ class JobController extends Controller
 
     public function update(Job $job)
     {
-        // Authorize (TODO)
+        Gate::authorize('edit-job', $job);
 
         // Validate
         request()->validate([
@@ -92,7 +85,7 @@ class JobController extends Controller
 
     public function destroy(Job $job)
     {
-        // Authorize (TODO)
+        Gate::authorize('edit-job', $job);
 
         // Delete
         $job->delete();
